requirements.txt
django==1.10
django-ckeditor==4.0.2
djangorestframework==3.4.7
flake8==3.0.4
isort==4.2.4
